This navigation_edges works almost right
This was my testing function, I just added proper math to it. I used it to make a .csv file, but a few of the costs are off from the example .csv, and as I understand it they should be identical. The shortest path it calculates is also not quite identical, though it's only different by one cell.